+++ app/src/main/java/io/github/christianjann/gittasks/ui/component/PickFolderDialog.kt
@@ -28,6 +28,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
+import io.github.christianjann.gittasks.ui.model.DrawerMode
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.lifecycle.ViewModel
@@ -178,8 +179,8 @@ private fun PickFolderDialogInternal(
             currentPath = currentNoteFolderRelativePath,
             openFolder = openFolder,
             createNoteFolder = createNoteFolder,
-            showTags = false,
-            onToggleMode = {},
+            drawerMode = DrawerMode.Folders,
+            onCycleMode = {},
             noteBeingMoved = null,
             onMoveNoteToFolder = {},
             onCancelMove = {},